+++ hois_back/src/main/java/ee/hitsa/ois/web/dto/PracticeJournalDto.java
@@ -19,7 +19,7 @@ public class PracticeJournalDto extends VersionedCommand {
     private AutocompleteResult module;
     private AutocompleteResult theme;
     private BigDecimal credits;
-    private Integer hours;
+    private Short hours;
     private LocalDate startDate;
     private LocalDate endDate;
     private String practicePlace;
@@ -114,11 +114,11 @@ public class PracticeJournalDto extends VersionedCommand {
         this.credits = credits;
     }
 
-    public Integer getHours() {
+    public Short getHours() {
         return hours;
     }
 
-    public void setHours(Integer hours) {
+    public void setHours(Short hours) {
         this.hours = hours;
     }
