+++ hois_back/src/main/java/ee/hitsa/ois/web/dto/curriculum/CurriculumVersionOccupationModuleCapacityDto.java
@@ -22,7 +22,7 @@ public class CurriculumVersionOccupationModuleCapacityDto extends VersionedComma
     @NotNull
     @Min(0)
     @Max(10000)
-    private Integer hours;
+    private Short hours;
 
     private Boolean contact = Boolean.FALSE;
 
@@ -48,11 +48,11 @@ public class CurriculumVersionOccupationModuleCapacityDto extends VersionedComma
         this.capacityType = capacityType;
     }
 
-    public Integer getHours() {
+    public Short getHours() {
         return hours;
     }
 
-    public void setHours(Integer hours) {
+    public void setHours(Short hours) {
         this.hours = hours;
     }
